Revert "Moves the config to a directory"
This reverts commit c38096ca45c6798a136229b30b3defa5687b299a.
