+++ core/actions/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const config = require('../../config/config');
+const config = require('../../config');
 
 const checkers = [
   require('../services/checkers/qa-review'),
